requirements.txt
disnake>=2.10.1
aiosqlite>=0.20.0
python-dotenv>=1.0.1
